Django-Angularjs-Fibonacci-Series-Generator-master/requirements.txt
Django==1.10.3
django-cors-headers==1.3.1
django-filter==1.0.1
djangorestframework==3.5.3
pytz==2017.3
virtualenv==15.1.0
virtualenvwrapper-win==1.2.5
